sp23-chips/vlsi/hammer-bump-plugin/pyproject.toml
[tool.poetry]
name = "intech_project"
version = "0.0.1"
description = "Hammer intech project example."
authors = ["Ken Ho <ken_ho@berkeley.edu>"]
maintainers = ["Ken Ho <ken_ho@berkeley.edu>"]
readme = "README.md"
repository = "https://github.com/ucb-bar/hammer"

[tool.poetry.dependencies]
python = "^3.9"
hammer-cadence-plugins = {path = "hammer-cadence-plugins", develop = true}
hammer-synopsys-plugins = {path = "hammer-synopsys-plugins", develop = true}
hammer-mentor-plugins = {path = "hammer-mentor-plugins", develop = true}
hammer-intech22-plugin = {path = "hammer-intech22-plugin", develop = true}
sty = "^1.0.4"
pytest = "^7.1"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.8", "setuptools>=65.3"]
build-backend = "poetry.core.masonry.api"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py39,py310
isolated_build = True

[testenv]
deps = pytest
commands =
  pytest tests/ {posargs}
"""
